==> notebook/1-2 Coleta de Dados - Concate_All_Csv_By_Scrapping.ipynb <==
# 1.2 Contat All CSV by Scrapping

import pandas as pd
import numpy as np
import os
import sys
sys.path.append('..')

# Import das extrações em csv através de scrapping:

guaruja1 = pd.read_csv('../data/data_scrapping/airbnb_guaruja_fev.csv')
guaruja2= pd.read_csv('../data/data_scrapping/airbnb_guaruja_mar.csv')
praiagrande1 = pd.read_csv('../data/data_scrapping/airbnb_praiagrande_fev.csv')
praiagrande2 = pd.read_csv('../data/data_scrapping/airbnb_praiagrande_mar.csv')
santos1 = pd.read_csv('../data/data_scrapping/airbnb_santos_fevereiro.csv')
santos2 = pd.read_csv('../data/data_scrapping/airbnb_santos_marco_.csv')
santos3 = pd.read_csv('../data/data_scrapping/airbnb_santos_abril_.csv')
santos4 = pd.read_csv('../data/data_scrapping/airbnbsantos_extracao_24012021_com_id.csv')
mongagua1 = pd.read_csv('../data/data_scrapping/airbnb_monguagua_fev.csv')
mongagua2 = pd.read_csv('../data/data_scrapping/airbnb_monguagua_mar.csv')
bertioga1 = pd.read_csv('../data/data_scrapping/airbnb_bertioga_fev.csv')
bertioga2 = pd.read_csv('../data/data_scrapping/airbnb_bertioga_mar.csv')
itanhaem1 = pd.read_csv('../data/data_scrapping/airbnb_itanhaem_fev.csv')
itanhaem2 = pd.read_csv('../data/data_scrapping/airbnb_itanhaem_mar.csv')
peruibe1 = pd.read_csv('../data/data_scrapping/airbnb_peruibe_fev.csv')
peruibe2 = pd.read_csv('../data/data_scrapping/airbnb_peruibe_mar.csv')

# Concat nos DataFrames
df_baixada = pd.concat([guaruja1, 
                                guaruja2, 
                                praiagrande1,
                                praiagrande2,
                                santos1,
                                santos2,
                                santos3,
                                santos4,
                                mongagua1,
                                mongagua2,
                                bertioga1,
                               bertioga2,
                               itanhaem1,
                               itanhaem2,
                               peruibe1,
                               peruibe2], ignore_index=True)

df_baixada.shape

df_baixada = df_baixada.drop_duplicates(keep='first')

df_baixada.head()

# Correção de erros da extração via Scrapping
df_baixada['check-out'] = df_baixada['check-out'].replace('previous_page_section_name=1000', '2021-05-02')

df_baixada['check-out'].unique()

# 1.2.1 - Função Limpeza Scrapping

def limpeza (dataset):
    """
    Funccion of cleaness
    dataset: Name of dataframe extracted from AirBnb's scraping
    """
    #drop da coluna unnamed
    dataset.drop(['Unnamed: 0'], axis=1, inplace=True)
    
    #rename columns:
    dataset = dataset.rename(columns={'id': 'ID', 
                                      'check-in': 'Check-In', 
                                      'check-out': 'Check-Out',
                                       'titulo': 'Localização', 
                                      'atributo_1': 'Número Hóspedes',
                                       'atributo_2': 'Número Quartos', 
                                      'atributo_3': 'Número Camas',
                                       'atributo_4': 'Número Banheiros', 
                                      'atributo_5':'Plus 1',
                                      'atributo_6':'Plus 2',
                                       'atributo_7':'Plus 3',
                                      'atributo_8':'Plus 4', 
                                      'avaliação':'Avaliação',
                                        'comentários': 'Número Comentários', 
                                      'preço': 'Preço/Noite',
                                      'preço_total': 'Preço com taxas',
                                       'superhost': 'Superhost'})
    #Find out Neighborhood:
    dataset['Localização'] = dataset['Localização'].str.replace('Apartamento inteiro em ','')
    dataset['Localização'] = dataset['Localização'].str.replace('Condomínio inteiro em ','')
    dataset['Localização'] = dataset['Localização'].str.replace('Quarto inteiro em ','')
    dataset['Localização'] = dataset['Localização'].str.replace('Flat inteiro em ','')
    dataset['Localização'] = dataset['Localização'].str.replace('Loft inteiro em ','')
    dataset['Localização'] = dataset['Localização'].str.replace('Suíte de visitas inteira em ','')
    
    #Only numbers
    dataset['Número Hóspedes'] = dataset['Número Hóspedes'].str.replace(' hóspedes','')
    dataset['Número Quartos'] = dataset['Número Quartos'].str.replace(' quartos','')
    dataset['Número Quartos'] = dataset['Número Quartos'].str.replace(' quarto','')
    dataset['Número Camas'] = dataset['Número Camas'].str.replace(' camas','')
    dataset['Número Camas'] = dataset['Número Camas'].str.replace(' cama','')
    dataset['Número Banheiros'] = dataset['Número Banheiros'].str.replace(' banheiros','')
    dataset['Número Banheiros'] = dataset['Número Banheiros'].str.replace(' banheiro','')
    
    x = dataset.drop_duplicates('Localização')
    np.array(x['Localização'])
    
    # Datatime transforming
    dataset['ID'] = dataset['ID'].astype('object')
    dataset['Check-In'] = pd.to_datetime(dataset['Check-In'], format='%Y-%m-%d')
    dataset['Check-Out'] = pd.to_datetime(dataset['Check-Out'], format='%Y-%m-%d')
    
    return dataset 

df_baixada = limpeza(df_baixada)
df_baixada.head()

print(f" {df_baixada.duplicated().sum()} linhas duplicadas")

df_baixada = df_baixada.drop_duplicates(keep='first')
print(f"Total de {df_baixada.duplicated().sum()} linhas duplicadas")

len(df_baixada['ID'].unique())

print(f'Numero de linhas: {df_baixada.shape[0]}')
print(f'Numero de colunas: {df_baixada.shape[1]}')

# Segundo Teste
df_baixada['Preço/Noite'].sort_values()

# 1.2.3. Exportar CSV

# Exportar para csv
df_baixada.to_csv('../data/baixada_santista.csv')
