==> notebook/2-1 Limpeza de Dados - Primeira Limpeza.ipynb <==
# 2.1 Limpeza de dados

## 2.1 Primeira Limpeza

import pandas as pd
import numpy as np

dados = pd.read_csv('airbnbsantos_extracao_11122020.csv')
dados_id = pd.read_csv('airbnbsantos_extracao_24012021_com_id.csv')

dados_id.head()

dados_id.drop(['Unnamed: 0'], axis=1, inplace=True)

dados_id.drop(['titulo_2'], axis=1, inplace=True)

dados_id.drop(['preço_total'], axis=1, inplace=True)

**O que é esse super host? Precisamos manter?**

Vamos manter por enquanto

dados_id.head()

dados_id = dados_id.rename(columns={'id': 'ID', 'check-in': 'Check-In', 'check-out': 'Check-Out',
                           'titulo': 'Localização', 'atributo_1': 'Número Hóspedes',
                           'atributo_2': 'Número Quartos', 'atributo_3': 'Número Camas',
                           'atributo_4': 'Número Banheiros', 'atributo_5':'Plus 1','atributo_6':'Plus 2',
                           'atributo_7':'Plus 3','atributo_8':'Plus 4', 'avaliação':'Avaliação',
                           'comentários': 'Número Comentários', 'preço': 'Preço/Noite',
                           'superhost': 'Superhost'})

dados_id.head()

np.array(dados_id['Localização'])

dados_id['Localização'] = dados_id['Localização'].str.replace('Apartamento inteiro em ','')
dados_id['Localização'] = dados_id['Localização'].str.replace('Condomínio inteiro em ','')
dados_id['Localização'] = dados_id['Localização'].str.replace('Quarto inteiro em ','')
dados_id['Localização'] = dados_id['Localização'].str.replace('Flat inteiro em ','')
dados_id['Localização'] = dados_id['Localização'].str.replace('Loft inteiro em ','')
dados_id['Localização'] = dados_id['Localização'].str.replace('Suíte de visitas inteira em ','')
dados_id['Número Hóspedes'] = dados_id['Número Hóspedes'].str.replace(' hóspedes','')
dados_id['Número Quartos'] = dados_id['Número Quartos'].str.replace(' quartos','')
dados_id['Número Quartos'] = dados_id['Número Quartos'].str.replace(' quarto','')
dados_id['Número Camas'] = dados_id['Número Camas'].str.replace(' camas','')
dados_id['Número Camas'] = dados_id['Número Camas'].str.replace(' cama','')
dados_id['Número Banheiros'] = dados_id['Número Banheiros'].str.replace(' banheiros','')
dados_id['Número Banheiros'] = dados_id['Número Banheiros'].str.replace(' banheiro','')

**Estamos tratando Número Banheiros / Camas / Quartos / Hóspedes como categóricas, pois temos estúdios, compartilhados.**

x = dados_id.drop_duplicates('Localização')
np.array(x['Localização'])

np.array(dados_id['Número Quartos'])

dados_id.isnull().sum()

dados_id.info()

dados_id['ID'] = dados_id['ID'].astype('object')
dados_id['Check-In'] = pd.to_datetime(dados_id['Check-In'], format='%Y-%m-%d')
dados_id['Check-Out'] = pd.to_datetime(dados_id['Check-Out'], format='%Y-%m-%d')

dados_id.info()

(dados_id['Check-In'] > '2021-01-31').sum()

(dados_id['Check-Out'] > '2021-01-31').sum()

dados_id.describe()

dados_id.shape

dados_id = dados_id.drop_duplicates(keep='first')

dados_id.shape

dados_id['ID'].unique()

colunas_constantes = []
for col in dados_id.columns:
    if len(dados_id[col].unique()) == 1:
        colunas_constantes.append(col)
    else:
        print(f'{col} OK')

colunas_constantes

dados_id['Check-In']

dados_id

dados_id.to_csv('Dados Airbnb Santos.csv')
