==> notebook/2-2 Limpeza de Dados - Segunda Limpeza.ipynb <==
# 2.2 Limpeza de Dados

### Segunda Limpeza

## 2.2.1 - Tratamento de Erros Scrapping

# Bibliotecas
import pandas as pd
import numpy as np

## Dados Constantes

DATA_INPUT_RAW = 'baixada_santista.csv'

df_raw = pd.read_csv(DATA_INPUT_RAW)

df_raw.head()

df_raw.shape

### Tratamento de Erros do Scrapping

Por algum motivo os valores de preço vieram com erros a partir do Scrapping. O era para ser 1022.0 está 1.022.

Neste caso, vamos fazer a correção desses valores.

As variáveis que estão com este problema são: **Preço/Noite** e **Preço com taxas**

print(df_raw['Preço/Noite'].sort_values())

# Preço/Noite:
print(len(df_raw.loc[df_raw['Preço/Noite'] < 10.0]))

# Preço com taxas
print(len(df_raw.loc[df_raw['Preço com taxas'] < 10.0]))

### Preço/Noite

filter_1 = df_raw.loc[(df_raw['Preço/Noite'] < 10.0), :].index

filter_1

df_raw.iloc[filter_1]['Preço/Noite'] * 1000

df_raw['Novo_Preço/Noite'] = df_raw.iloc[filter_1]['Preço/Noite'] * 1000

values = df_raw['Preço/Noite']

df_raw['Novo_Preço/Noite'] = df_raw['Novo_Preço/Noite'].fillna(values)

df_raw.head()

**Verificação:**

# Preço/Noite:
print(len(df_raw.loc[df_raw['Novo_Preço/Noite'] < 10.0]))

### Preço com taxas

filter_2 = df_raw.loc[(df_raw['Preço com taxas'] < 10.0), :].index

filter_2

df_raw['Novo_Preço com taxas'] = df_raw.iloc[filter_2]['Preço com taxas'] * 1000

value_2 = df_raw['Preço com taxas']

df_raw['Novo_Preço com taxas'] = df_raw['Novo_Preço com taxas'].fillna(value_2)

df_raw.head()

**Verificação**

# Preço com taxas:
df_raw.loc[df_raw['Novo_Preço com taxas'] < 10.0]

Os erros foram tratados.

Agora vamos dropar as variáveis com erro  renomea-las.

df_raw = df_raw.drop({'Preço/Noite', 'Preço com taxas'}, axis=1)

df_raw.rename(columns = {'Novo_Preço/Noite': 'Preço/Noite', 
                                      'Novo_Preço com taxas': 'Preço com taxas'}, inplace=True)

## Tratamento finalizado

df_raw

# 2.2.2 Exportar

df_raw.to_csv('DATA_baixada_santista_raw.csv', index=False)

# 2.2.3 Função de Tratamento

df_test_funccion = pd.read_csv('baixada_santista.csv')

def corrigir_monetario(df, name='name'):
    """
    Correção de valores float incorretos 1.0 para 1000.0
    df = Dataframe
    name= Nome da variável    
    """
    
    filter_1 = df.loc[(df[name] < 10.0), :].index   
    df['New_name'] = df.iloc[filter_1][name] * 1000    
    value = df[name]    
    df['New_name'] = df['New_name'].fillna(value)
    df = df.drop(name, axis=1)
    df.rename(columns = {'New_name': name}, inplace=True)
    
    return df

df_test_funccion = corrigir_monetario(df_test_funccion, name='Preço/Noite')

df_test_funccion = corrigir_monetario(df_test_funccion, name='Preço com taxas')

df_test_funccion.head()
