==> notebook/2-3 Limpeza de Dados - Terceira Limpeza.ipynb <==
# 2.3 Limpeza de Dados

### Terceira Limpeza

# Biblioteca
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
pd.set_option('display.max_columns',60)

## Dados Constantes

DATA_OUTPUT_RAW = '../data/DATA_baixada_santista_raw.csv'
DATA_OUTPUT_TRAIN_CLEANED = '../data/DATA_train_cleaned.csv'
DATA_TEST = '../data/DATA_TEST.csv'
TARGET = 'Preço/Noite'

df_raw = pd.read_csv(DATA_OUTPUT_RAW)

df_raw.head()

df_raw.shape

### Linhas duplicadas

print(f' Linhas duplicadas: {df_raw.duplicated().sum()}')

### Colunas duplicadas

df_raw_T = df_raw.T
df_raw_T[df_raw_T.duplicated(keep=False)]

### Colunas Constantes

columns_to_drop_constant = []
for col in df_raw.columns:
    if df_raw[col].nunique() == 1:
        columns_to_drop_constant.append(col)
    else:
        continue

columns_to_drop_constant

## 2.3.1. Divisão dos Dados

df_train, df_test = train_test_split(df_raw,
                                     test_size=0.2,
                                     random_state=123)
df_train.shape, df_test.shape

## 2.3.2 Damificação de variáveis

### Criação de Dummies

Ainda temos que criar uma função que coloque no For todos os novas palavras que possam vir com atualizações do dataset. Por hora essa lista está estática, o ideal é que ela seja dinâmica

def create_dummies(df):
    """
    Create dummies from Plus and Superhost columns
    df = DataFrame from airbnb scrapping
    Return DataFrame with dummies columns and dropped the originals
    """ 
    # Dummies das colunas Plus
    plus1 = df['Plus 1'].unique()
    plus2 = df['Plus 2'].unique()
    plus3 = df['Plus 3'].unique()
    plus4 = df['Plus 3'].unique()
    fullplus = plus1.tolist() + plus2.tolist() + plus3.tolist() + plus4.tolist()
    fullplus = set(fullplus)
    def word_in_columns(df, word):
        if word in df['Plus 1'] or word in df['Plus 2'] or word in df['Plus 3'] or word in df['Plus 4']:
            return 1
        else:
            return 0
        
    count = 0
    for word in fullplus:
        df[f'{word}'] = df.apply(word_in_columns, axis=1, args= (word, ))
        count += 1
    
    # Dummies para Superhost
    Superhost = df['Superhost'].unique()
    def host_in_columns(df, word):
        if word in df['Superhost']:
            return 1
        else:
            return 0
        
    count = 0
    for word in Superhost:
        df[f'{word}'] = df.apply(word_in_columns, axis=1, args= (word, ))
        count += 1
    
    #Drop das colunas 
    df = df.drop(columns=['Plus 1' , 'Plus 2', 'Plus 3', 'Plus 4', 'Superhost'])
    
    return df

df_train = create_dummies(df_train)

df_train.head()

### Tratamento do Número de Quartos

def quartos_transformation(df):
    df.replace({'Número Quartos' : {'Estúdio': '1'}}, inplace=True)
    return df

df_train = quartos_transformation(df_train)

### Tratamento do Número de banheiros

# verificando as inconsistências:

df_train['Número Banheiros'].value_counts()

**Função de tratamento de problemas**

def banheiro_transformation(df):
    """
    Extract any word inside Número de banheiros column and create a new one called Banheiro Compartilhado.
    In banheiro Compartilhado: If True, There is Banheiro compartilhado. If False, There is no Banheiro compartilhado.
    df = raw dataset cleanned from airbnb scrapping
    """
    # Pega o qualquer texto após o espaço.
    df['Banheiro Compartilhado'] = df['Número Banheiros'].str.contains('\\ ', regex=True)
    
    # Pega o primeiro digito numérico
    df['Número Banheiros'] = df['Número Banheiros'].str[0:1]
    df['Número Banheiros'] = df['Número Banheiros'].str.strip()
    
    return df

df_train = banheiro_transformation(df_train)

df_train['Número Banheiros'].value_counts()

## 2.3.2 Tratamento de variáveis

# Vamos dropar a coluna Unnamed: 0 e  Titulo_2
df_train = df_train.drop({'Unnamed: 0', 'titulo_2'}, axis= 1)

# Análise de Variáveis
df_train.dtypes

### Ajuste dos Tipos

# A variável número de quartos e número de banheiros estavam como categóricas. Então, fez se necessário a conversão para numérica

# Transformação em datatime as variáveis Check-in e Check-out:

def ajuste_tipos(df):
    """
    Types transformation
    df= dataset
    """
    
    # Object to Numeric
    df['Número Quartos'] = pd.to_numeric(df['Número Quartos'], errors = 'coerce')
    df['Número Banheiros'] = pd.to_numeric(df['Número Banheiros'], errors = 'coerce')
    # To DateTime
    df['Check-In'] = pd.to_datetime(df_train['Check-In'], format ='%Y-%m-%d')
    df['Check-Out'] = pd.to_datetime(df_train['Check-Out'], format ='%Y-%m-%d')
    
    return df

df_train = ajuste_tipos(df_train)

# Agora todos os tipos estão classificados corretamente.
df_train.dtypes

# Checando valores atípicos
df_train.describe().T

Valores atítpicos identificados: <br>
Há quartos em que não 0 camas e 0 Banheiros.

# Invetigando o número de camas.
df_train['Número Camas'].value_counts()

# Supondo que se há quarto é possivel que haja cama.
len(df_train.loc[(df_train['Número Camas'] == 0 & (df_train['Número Quartos'] >=1))])

Entretanto, Existe a possibilidade do locatário alugar o imóvel por temporada sem disponibilizar uma cama.<br>
Pesando nesta linha, vamos optar por não alterar e/ou remover estes imóveis

# Exportar

df_train.to_csv(DATA_OUTPUT_TRAIN_CLEANED)
df_test.to_csv(DATA_TEST)
