==> notebook/4 Feature_selection.ipynb <==
# 4. Feature Selection

# libs:
import pandas as pd
import numpy as np
pd.set_option('display.max_columns',60)

#Testes estatísticos
import scipy
from scipy.stats import stats
import statsmodels.api as sm

from feature_engine.categorical_encoders import CountFrequencyCategoricalEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

#fearture selection:
from sklearn.feature_selection import chi2
from statsmodels.stats.outliers_influence import variance_inflation_factor
from scipy.stats import chi2_contingency
from boruta import BorutaPy
from sklearn.feature_selection import RFECV
from sklearn.ensemble import RandomForestRegressor

import warnings
warnings.filterwarnings("ignore")

Vamos utilizar o método de votação marjoritáia. 3 Propostas de seleção.
<br>
* Teste estatístico
* RFECV
* Boruta<br>

As features que forem mantidas por pelo menos 2 algoritmos serão utilizadas no processo de modelagem

## Dados Constantes

DATA_INPUT_ENG = '../data/DATA_train_eng.csv'
DATA_OUTPUT_FEAT_SELECTED = '../data/DATA_feature_selected.csv'

df_train = pd.read_csv(DATA_INPUT_ENG)

**Dropar variáveis que não vamos utilizar**

# Vamos dropar algumas colunas
df_train = df_train.drop({'Unnamed: 0','ID', 'Check-In', 'Check-Out', 'Preço com taxas', 'Diária com taxas'}, axis=1)

df_train.head()

df_train.shape

# Divisão para os testes Wrap methods
X_train = df_train.drop('Preço/Noite', axis=1)
y_train = df_train['Preço/Noite']

**Tratamento de alguns tipos:**

df_train.dtypes

**Divisão entre as variáveis númericas e categóricas**

features_num = ['Número Hóspedes', 'Número Quartos', 'Número Camas','Número Banheiros', 
                 'Avaliação', 'Número Comentários', 'Taxa', 'Dia', 'Mes']

features_cat = [ 'Localização', 'Academia','Ar-condicionado', 'Cozinha', 'Elevador',
                       'Estacionamento gratuito', 'Máquina de Lavar', 'Permitido animais',
                       'Piscina', 'Secadora', 'Self check-In', 'Wi-Fi', 'Café da Manhã',
                       'Entrada/saída para esquis', 'Jacuzzi', 'Lareira interna',
                       'Lava-louças', 'Novo preço mais baixo', 'Raridade',
                       'Banheiro Compartilhado','Vista', 'Demanda', 'É_feriado',
                       'É outilier', 'Semana_do_ano', 'Semana de Feriado']

Target = ['Preço/Noite']

# Conversão de Tipos
for col in features_cat:
    df_train[col] = df_train[col].astype(str)

df_train.dtypes

Tudo Pronto!

## 4.1 Testes Estatísticos:

### Features Númericas:

### 1. Teste de Correlação

# Teste de correlação
def teste_correlacao(df, Target, features, alpha=0.05):
    """
    df = Dataframe
    Target = target
    features = numeric features to be tested
    Return list of variable which are correlationed with target
    """
    correlation = []
    no_correlation = []
    for col in features:
        cor, p = stats.spearmanr(df_train[col],df_train[Target])
        if p <= alpha:
            correlation.append(col)
        else:
            no_correlation.append(col)
            
    return correlation

correlation = teste_correlacao(df_train, Target, features_num)
correlation

### Features categóricas:

### Chi2

def teste_chi2_(df, Target='Target', features_cat=features_cat, aplha=0.05):
    """
    Chi2 test
    df = DataFrame
    Target = Target string
    features = List of categorical features
    aplha = indice of significance
    Return a list of variable which has passed for Chi2 test
    """
    from scipy.stats import chi2_contingency
    p_values_cat_features = {}
    for col in features_cat:
        # Cria tabela de contingencia
        df_cross = pd.crosstab(df[col], df[Target])
        # Aplica o teste e extrai o p-valor
        p_value = scipy.stats.chi2_contingency(df_cross)[1]
        # Armazena coluna e p-valor em um dict
        p_values_cat_features[col] = p_value
        
    p_values_cat_features = pd.Series(p_values_cat_features)
    filter_cat_features = p_values_cat_features[p_values_cat_features < aplha].index.tolist()
        
    return filter_cat_features

chi2 = teste_chi2_(df_train, Target = 'Preço/Noite', features_cat=features_cat)

chi2

Pelo processo de correlação, todas passaram pelo teste, porém pelo teste de Multicolinearidade. Número de quartos e número de hospedes não passaram.

## 4.2 RFECV

Para este teste vamos utilizar o Dataframe encode

def teste_rfe(X_train, y_train):
    """
    RFE Test
    X_train = Dataframe with all features to be tested and not target
    y_train = Target
    Return a list of varibales which has passed for the test
    """
    from sklearn.feature_selection import RFECV
    from sklearn.ensemble import RandomForestRegressor
    rfecv_RFC = RFECV(estimator=RandomForestRegressor(n_jobs = -1, max_depth = 5), scoring='neg_mean_squared_error')
    rfecv_RFC.fit(X_train,y_train)
    mask_RFC = rfecv_RFC.support_
    cols_selected_RFE = X_train.columns.tolist()
    
    return cols_selected_RFE

rfe = teste_rfe(X_train, y_train)

rfe

## 4.3 Boruta

def teste_boruta(X_train, y_train):
    """
    Boruta Test - Return features which was aproved
    X_train = Dataframe with all features to be tested and not target
    y_train = Target
    Return a list of varibales which has passed for the test
    """
    from boruta import BorutaPy
    from sklearn.ensemble import RandomForestRegressor
    # -1 indica para o python usar todo o processador 
    boruta_selector = BorutaPy(RandomForestRegressor(n_jobs = -1, max_depth = 5), n_estimators = 50, max_iter=100, random_state = 0)
    boruta_selector.fit(np.array(X_train), np.array(y_train))
    boruta_selector.support_
    boruta_selector.support_weak_
    features_boruta = X_train.loc[:, boruta_selector.support_].columns.tolist()

    return features_boruta 

boruta = teste_boruta(X_train, y_train)

boruta

# 4.4 Variáveis Selecionadas

### Testes estatísticos:

correlation

chi2

#vif

#statistic_test_num = list(set(correlation).intersection(set(vif)))

def feature_results___(correlation, chi2, rfe, boruta):
    
    # Junção dos testes de correlação e Chi2
    statistic_test = correlation + chi2
    # Interserção entre testes estatísticos e RFE
    statistic_test_rfe = list(set(statistic_test).intersection(set(rfe)))
    # Quem passou no primeiro que também passou no boruta
    feature_selection = list(set(statistic_test_rfe).intersection(set(boruta)))
    
    return feature_selection

def feature_results(correlation, chi2, rfe, boruta):
    """
    Return a list which has passed for three testes
    correlation = set of result from corrrelation teste
    chi2 = set of features which has passed from chi2 teste
    rfe = set of features which has passed from rfe teste
    boruta = set of features which has passed from boruta teste
    """
    # Junção dos testes de correlação e Chi2
    statistic_test = correlation + chi2
    # Interserção entre testes estatísticos e RFE
    statistic_test_rfe = list(set(statistic_test).intersection(set(rfe)))
    # Interserção entre testes estatísticos e boruta
    statistic_test_boruta = list(set(statistic_test).intersection(set(boruta)))
    # Interserção entre rfe e boruta
    rfe_boruta = list(set(rfe).intersection(set(boruta)))
    
    # Quem passou no primeiro que também passou no boruta
    feature_selection = set(statistic_test_rfe) & set(statistic_test_boruta) & set(rfe_boruta)
    
    return feature_selection

feature_results(correlation, chi2, rfe, boruta)

feature_selection = feature_results(correlation, chi2, rfe, boruta)

feature_selection

feature_selection_num = []
feature_selection_cat = []
count = 0
for i in feature_selection:
    if i in features_num:
        feature_selection_num.append(i)
    if i in features_cat:
        feature_selection_cat.append(i)
        count += 1

feature_selection_num

feature_selection_cat

# 4.5 Teste de Multicolineariade

def multicolinearidade(df, features):
    """
    V-Cramer test
    X = DataFrame without target
    features = numerical features to be tested
    Return a Dataframe with VIF score
    """  
    # VIF dataframe 
    vif_data = pd.DataFrame() 
    vif_data["feature"] = df.loc[:,features].columns 
    # calculando VIF 
    vif_data["VIF"] = [variance_inflation_factor(df.loc[:,features].values, i) for i in range(len(df.loc[:,features].columns))] 
    
    return vif_data.sort_values(by='VIF', ascending=False)

multicolinearidade(X_train, feature_selection_num)

feature_selection_num.remove('Avaliação')

multicolinearidade(X_train, feature_selection_num)

feature_selection_num.remove('Número Quartos')

multicolinearidade(X_train, feature_selection_num)

vif = multicolinearidade(X_train, feature_selection_num)

def vif_list(vif):
    """
    Tranforming resut VIF test into a list of variables
    vif = DataFrame from teste_vif
    Return a list of variables 
    """
    vif = vif.pivot_table(columns='feature')
    vif = vif.columns.tolist()
    
    return vif

vif = vif_list(vif)

# Variáveis sem multicolinearidade
vif

# Agora vamos juntar estas variáveis áquela que dropamos ante de passar no VIF.

Finalizado todos os testes, Temos um conjunto sem multicolinearidade para o modelo.

df_train = df_train[vif + feature_selection_cat + Target]

df_train

print(len(df_train.columns))

# criar função de junção das features selecionadas no dataframe

# Export

DATA_OUTPUT_FEAT_SELECTED_V2 = '../data/DATA_feature_selected_V2.csv'

df_train.to_csv(DATA_OUTPUT_FEAT_SELECTED_V2)
